Add files via upload
This is a modern adaptation of the classic Pong game for Android devices. Built with intuitive touch controls, the game delivers an exciting table-tennis experience with a minimalistic yet engaging design. Players use swipes to control their paddles, bouncing the ball back and forth in a fast-paced match.
key Features:
1) single Player
2)two player mode
3) touch control
4) Light weight and Smooth